Remove more redundancy
